+++ src/packages/piral-cli/src/apps/build-piral.ts
@@ -5,15 +5,21 @@ import { extendConfig, setStandardEnvs, checkExists, findFile, getPiletsInfo } f
 export interface BuildPiralOptions {
   entry?: string;
   target?: string;
+  publicUrl?: string;
 }
 
 export const buildPiralDefaults = {
   entry: './src/index.html',
   target: './dist/index.html',
+  publicUrl: '/',
 };
 
 export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOptions = {}) {
-  const { entry = buildPiralDefaults.entry, target = buildPiralDefaults.target } = options;
+  const {
+    entry = buildPiralDefaults.entry,
+    target = buildPiralDefaults.target,
+    publicUrl = buildPiralDefaults.publicUrl,
+  } = options;
   const entryFiles = join(baseDir, entry);
   const exists = await checkExists(entryFiles);
 
@@ -44,6 +50,7 @@ export async function buildPiral(baseDir = process.cwd(), options: BuildPiralOpt
       minify: true,
       scopeHoist: false,
       contentHash: true,
+      publicUrl,
     }),
   );